requirements.txt
ipykernel
pandas
fastapi
sqlalchemy
python-dotenv
uvicorn
psycopg[binary]
python-jose[cryptography]
bcrypt
python-multipart
pyjwt
-e .